requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.5,<2.8.6
markdown>=3.2.2,<3.2.3
